Remove duplicate media lib and consolidate to src/scripts/media.
