+++ packages/openneuro-app/src/scripts/refactor_2021/dataset/files/file.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
 import UpdateFile from '../mutations/update-file.jsx'
 import DeleteFile from '../mutations/delete-file.jsx'
-import { Media } from '../styles/media'
+import { Media } from '../../../styles/media'
 import RemoveAnnexObject from '../mutations/remove-annex-object.jsx'
 import FlagAnnexObject from '../mutations/flag-annex-object.jsx'
 import { isAdmin } from '../../authentication/admin-user.jsx'
