--- packages/openneuro-app/src/scripts/refactor_2021/dataset/styles/media.tsx
@@ -1,12 +1,0 @@
-import { createMedia } from '@artsy/fresnel'
-
-const AppMedia = createMedia({
-  breakpoints: {
-    small: 0,
-    medium: 768,
-  },
-})
-
-// Generate CSS to be injected into the head
-export const mediaStyle = AppMedia.createMediaStyle()
-export const { Media, MediaContextProvider } = AppMedia
